refactor(bit_buffer): parameterize BitBuffer on B<:Unsigned
Step 1 of the sync-detection redesign. Threads a `B<:Unsigned` type
parameter through `BitBuffer` and `TrackedSignal{Sig, B, C, PCF}` so
that the sync-search sliding window can be sized per signal without
disturbing the post-sync decoded-bits buffer (still `UInt128`).

Default `BitBuffer()` continues to return `BitBuffer{UInt128}`, so
behavior is unchanged at this step. Step 2 introduces the per-signal
`get_code_block_buffer_type` trait that lets each `TrackedSignal` pick
a narrower or wider buffer.

The existing `test/bit_buffer.jl` "Find bit start and buffer bit" case
relied on implicit promotion of a UInt64 literal into a UInt128 field;
with the field now parametric the literal stays UInt64 and the test's
expected 65-bit post-shift value gets truncated. Pinning the literal
to UInt128 restores the original semantics and documents the design
contract.

+++ test/bit_buffer.jl
@@ -40,7 +40,11 @@ using Tracking: BitBuffer, buffer, has_bit_or_secondary_code_been_found
     end
 
     @testset "Find bit start and buffer bit" begin
-        code_blocks_buffer = 0xfffffffffff80000
+        # `UInt128` literal so that the expected post-shift value
+        # 0x1fffffffffff00000 (65 bits) fits without truncation. The
+        # buffer width is now a per-`BitBuffer{B}` parameter, so the
+        # test pins `B = UInt128` explicitly.
+        code_blocks_buffer = UInt128(0xfffffffffff80000)
         code_blocks_buffer_length = ndigits(code_blocks_buffer; base = 2)
         bit_buffer = BitBuffer(
             code_blocks_buffer,
